实现天气界面的显示

+++ app/src/main/java/com/zhongyang/xiangyangweather/logic/model/RealtimeResponse.kt
@@ -43,7 +43,7 @@ data class RealtimeResponse(
 
     data class AirQuality(
         val aqi: Aqi,
-        val co: Int,
+        val co: Double,
         val description: Description,
         val no2: Int,
         val o3: Int,
